refactor: divide utils into fileUtils and utils
