+++ packages/botonic-plugin-nlu/package.json
@@ -26,8 +26,7 @@
     "fs": "0.0.1-security",
     "inquirer": "^6.3.1",
     "node-fetch": "^2.6.0",
-    "sqlite-async": "^1.0.11",
-    "mkdirp": "^0.5.1"
+    "sqlite-async": "^1.0.11"
   },
   "devDependencies": {
     "@babel/cli": "^7.5.5",
